get hp page in common setup
